+++ tests/testthat/test-slice.r
@@ -113,22 +113,36 @@ test_that("get_slice_size() validates its inputs", {
   })
 })
 
-test_that("get_slice_size() converts proportions to numbers", {
-  expect_equal(get_slice_size(prop = 0.5)(10), 5)
-  expect_equal(get_slice_size(prop = -0.1)(10), 9)
-  expect_equal(get_slice_size(prop = 1.1)(10), 11)
+test_that("get_slice_size() standardises prop", {
+  expect_equal(get_slice_size(prop = 0)(10), 0)
+
+  expect_equal(get_slice_size(prop = 0.4)(10), 4)
+  expect_equal(get_slice_size(prop = 2)(10), 10)
+  expect_equal(get_slice_size(prop = 2, allow_outsize = TRUE)(10), 20)
+
+  expect_equal(get_slice_size(prop = -0.4)(10), 6)
+  expect_equal(get_slice_size(prop = -2)(10), 0)
 })
 
-test_that("get_slice_size() converts negative to positive", {
-  expect_equal(get_slice_size(n = -1)(10), 9)
-  expect_equal(get_slice_size(prop = -0.5)(10), 5)
+test_that("get_slice_size() standardises n", {
+  expect_equal(get_slice_size(n = 0)(10), 0)
+
+  expect_equal(get_slice_size(n = 4)(10), 4)
+  expect_equal(get_slice_size(n = 20)(10), 10)
+  expect_equal(get_slice_size(n = 20, allow_outsize = TRUE)(10), 20)
+
+  expect_equal(get_slice_size(n = -4)(10), 6)
+  expect_equal(get_slice_size(n = -20)(10), 0)
 })
 
-test_that("get_slice_size() rounds non-integers", {
-  expect_equal(get_slice_size(n = 1.6)(10), 1)
+test_that("get_slice_size() rounds prop in the right direction", {
   expect_equal(get_slice_size(prop = 0.16)(10), 1)
-  expect_equal(get_slice_size(n = -1.6)(10), 9)
   expect_equal(get_slice_size(prop = -0.16)(10), 9)
+})
+
+test_that("n must be an integer", {
+  df <- tibble(x = 1:5)
+  expect_snapshot(slice_head(df, n = 1.1), error = TRUE)
 })
 
 test_that("functions silently truncate results", {
@@ -293,6 +307,12 @@ test_that("slice_sample() respects weight_by and replaces", {
   expect_equal(out$x, c(1, 1))
 })
 
+test_that("slice_sample() can increase rows iff replace = TRUE", {
+  df <- tibble(x = 1:10)
+  expect_equal(nrow(slice_sample(df, n = 20, replace = FALSE)), 10)
+  expect_equal(nrow(slice_sample(df, n = 20, replace = TRUE)), 20)
+})
+
 test_that("slice_sample() checks size of `weight_by=` (#5922)", {
   df <- tibble(x = 1:10)
   expect_snapshot(slice_sample(df, n = 2, weight_by = 1:6), error = TRUE)
@@ -312,16 +332,20 @@ test_that("`slice_sample()` validates `replace`", {
   })
 })
 
-test_that("slice_sample() handles n= and prop=", {
+test_that("slice_sample() handles positive n= and prop=", {
   gf <- group_by(tibble(a = 1, b = 1), a)
   expect_equal(slice_sample(gf, n = 3, replace = TRUE), gf[c(1, 1, 1), ])
   expect_equal(slice_sample(gf, prop = 3, replace = TRUE), gf[c(1, 1, 1), ])
+})
 
-  # Unlike other helpers, can't supply negative values
-  expect_snapshot(error = TRUE, {
-    slice_sample(gf, n = -1)
-    slice_sample(gf, prop = -1)
-  })
+test_that("slice_sample() handles negative n= and prop= (#6402)", {
+  df <- tibble(a = 1:2)
+  expect_equal(nrow(slice_sample(df, n = -1)), 1)
+  expect_equal(nrow(slice_sample(df, prop = -0.5)), 1)
+
+  # even if larger than n
+  expect_equal(nrow(slice_sample(df, n = -3)), 0)
+  expect_equal(nrow(slice_sample(df, prop = -2)), 0)
 })
 
 # slice_head/slice_tail ---------------------------------------------------